fix(stylelint): change `selector-pseudo-element-colon-notation` from single to double
Error:
- The `selector-pseudo-element-colon-notation` rule was set to `single`

Reason:
- Pseudo-elements should use double colon notation (`::before`, `::after`) as per the CSS3 specification

Solution:
- Change the rule value from `"single"` to `"double"` in `stylelint.config.js`

References:
- Issue #1107

+++ stylelint.config.js
@@ -156,7 +156,7 @@ module.exports = {
 				],
 			},
 		],
-		"selector-pseudo-element-colon-notation": "single",
+		"selector-pseudo-element-colon-notation": "double",
 		"selector-pseudo-element-no-unknown": [
 			true,
 			{
